simplify tax code
