+++ types/filecontracts.go
@@ -4,7 +4,6 @@ package types
 // contracts.
 
 import (
-	"github.com/NebulousLabs/Sia/build"
 	"github.com/NebulousLabs/Sia/crypto"
 )
 
@@ -112,11 +111,8 @@ func (fcid FileContractID) StorageProofOutputID(proofStatus ProofStatus, i uint6
 func (fc FileContract) Tax() Currency {
 	// COMPATv0.4.0 - until the first 20,000 blocks have been archived, they
 	// will need to be handled in a special way.
-	CurrentHeightLock.Lock()
-	height := CurrentHeight
-	CurrentHeightLock.Unlock()
-	if (height < 21e3 && build.Release == "standard") || (height < 10 && build.Release == "testing") {
-		return fc.Payout.MulFloat(0.039).RoundDown(SiafundCount)
-	}
+	// if (height < 21e3 && build.Release == "standard") || (height < 10 && build.Release == "testing") {
+	// return fc.Payout.MulFloat(0.039).RoundDown(SiafundCount)
+	// }
 	return fc.Payout.MulTax().RoundDown(SiafundCount)
 }
